Turn divide-and-conquer functionality into an extension.
