+++ test/runtests.jl
@@ -1,3 +1,4 @@
+using DecisionTree: Leaf, Node
 using HerbCore
 using HerbSearch
 using HerbGrammar
@@ -17,8 +18,6 @@ Random.seed!(1234)
 		piracies = (treat_as_own = [RuleNode, AbstractGrammar],),
 	)
 	include("test_search_procedure.jl")
-	include("test_divide_conquer.jl")
-	include("test_divide_conquer_example.jl")
 	include("test_context_free_iterators.jl")
 	include("test_sampling.jl")
 	include("test_stochastic/test_stochastic.jl")
@@ -32,6 +31,11 @@ Random.seed!(1234)
 	include("test_unique.jl")
 	include("test_constraints.jl")
 
-	# Excluded because it contains long tests
-	# include("test_realistic_searches.jl")
+	# 	# Excluded because it contains long tests
+	# 	# include("test_realistic_searches.jl")
+end
+
+@testset verbose = true "Divide and conquer extension" begin
+	include("test_divide_conquer.jl")
+	include("test_divide_conquer_example.jl")
 end
