+++ test/test_divide_conquer_example.jl
@@ -1,3 +1,6 @@
+DivideAndConquerExt = Base.get_extension(HerbSearch, :DivideAndConquerExt)
+using .DivideAndConquerExt: divide_and_conquer
+
 # Example problem, grammar bit functions taken from HerbBenchmarks: src/data/SyGuS/PBE_BV_Track_2018
 # 
 # HerbBenchmarks.jl is not a released package yet and can't be included as dependency
@@ -78,7 +81,7 @@ if0_cvc(x::UInt, y::UInt, z::UInt) = x == UInt(0) ? y : z
 
 	iterator = BFSIterator(grammar, :Start)
 	idx_ifelse = findfirst(r -> r == :($sym_bool ? $sym_start : $sym_start), grammar.rules)
-	@test_throws HerbSearch.ConditionalIfElseError HerbSearch.divide_and_conquer(
+	@test_throws DivideAndConquerExt.ConditionalIfElseError divide_and_conquer(
 		problem,
 		iterator,
 		sym_bool,
@@ -92,7 +95,7 @@ if0_cvc(x::UInt, y::UInt, z::UInt) = x == UInt(0) ? y : z
 	add_rule!(grammar, :($sym_start = $sym_bool ? $sym_start : $sym_start))
 	iterator = BFSIterator(grammar, :Start)
 
-	final_program = HerbSearch.divide_and_conquer(
+	final_program = divide_and_conquer(
 		problem,
 		iterator,
 		sym_bool,
